Moved reconutils.py in bnm.recon.algo directory and changed imports and usages from shell files.
