+++ bin/dev/seeg-ct-dw.sh
@@ -126,7 +126,7 @@ source snapshot.sh 3vols CT-reo.nii.gz CT-mask.nii.gz CT-dil-mask.nii.gz
 
 # label dil mask and apply to mask
 # XXX reported number of electrodes should match implantation schema
-python -c "import reconutils; reconutils.label_with_dilation('CT-mask.nii.gz', 'CT-dil-mask.nii.gz', 'CT-lab-mask.nii.gz')"
+python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.label_with_dilation('CT-mask.nii.gz', 'CT-dil-mask.nii.gz', 'CT-lab-mask.nii.gz')"
 
 #Visual checks (screenshot)
 # XXX check that labels have different colors, in mrview if possible
@@ -137,7 +137,7 @@ source snapshot.sh 2vols CT-reo.nii.gz CT-lab-mask.nii.gz
 
 # find contact positions
 python<<EOF
-import nibabel, numpy as np, reconutils
+import nibabel, numpy as np, bnm.recon.algo.reconutils
 nii = nibabel.load('CT-lab-mask.nii.gz')
 lab_bin = nii.get_data()
 aff = nii.affine
@@ -146,7 +146,7 @@ ulab = ulab[ulab > 0]
 ul_to_pos = {}
 all_pos = []
 for ul in ulab[ulab > 0]:
-    xyz_pos = reconutils.periodic_xyz_for_object(lab_bin, ul, aff)
+    xyz_pos = bnm.recon.algo.reconutils.periodic_xyz_for_object(lab_bin, ul, aff)
     all_pos.append(xyz_pos)
 all_pos = np.concatenate(all_pos, axis=0)
 np.savetxt('seeg_xyz.txt', all_pos, fmt='%f')
