+++ bin/segment.sh
@@ -155,7 +155,7 @@ then
     #Get the voxels that lie at the external surface (or border) of the target volume structures:
 
     #Using my python code:
-    #python -c "import reconutils; reconutils.vol_to_ext_surf_vol('$MRI/$vol.nii.gz',labels='$ASEG_LIST',hemi='lh rh',out_vol_path='./$vol-surf.nii.gz',labels_surf=None,labels_inner='0')"
+    #python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.vol_to_ext_surf_vol('$MRI/$vol.nii.gz',labels='$ASEG_LIST',hemi='lh rh',out_vol_path='./$vol-surf.nii.gz',labels_surf=None,labels_inner='0')"
     #flirt -applyxfm -in ./$vol-surf.nii -ref $DMR/b0.nii.gz -out ./$vol-surf-in-d.nii.gz -init $DMR/t2d.mat -interp nearestneighbour
 
     #Using freesurfer:
@@ -200,7 +200,7 @@ fi
 
 #Apply the mask to the parcellation volume now to get the parcellation volume of the regions that white matter touches. i.e., that survive the mask constructed above:
 #Using my python code:
-#python -c "import reconutils; reconutils.mask_to_vol('$mask_this_vol.nii.gz','./mask-$SEGMENT_METHOD.nii.gz','./$vol-mask.nii.gz',labels='$ASEG_LIST',hemi='lh rh',vol2mask_path=None,vn=$VOL_VN,th=1,labels_mask=None,labels_nomask='0')"
+#python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.mask_to_vol('$mask_this_vol.nii.gz','./mask-$SEGMENT_METHOD.nii.gz','./$vol-mask.nii.gz',labels='$ASEG_LIST',hemi='lh rh',vol2mask_path=None,vn=$VOL_VN,th=1,labels_mask=None,labels_nomask='0')"
 
 #Using freesurfer:
 mris_calc $mask_this_vol.nii.gz masked ./mask-$SEGMENT_METHOD.nii.gz
@@ -218,12 +218,12 @@ python -m $SNAPSHOT --snapshot_name $vol-mask 2vols $MRI/T1.nii.gz ./$vol-mask.n
 #Give an empty list for add_lbl, if you want cerebral white matter to be masked out
 for h in lh rh
 do
-python -c "import reconutils; reconutils.sample_vol_on_surf('$SURF/$h.white','./$vol-mask.nii.gz','$LABEL/$h.aparc.annot','./$h.white-mask','$CRAS_PATH',ctx='$h',vn=$SURF_VN,add_lbl=[2,41])"
+python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.sample_vol_on_surf('$SURF/$h.white','./$vol-mask.nii.gz','$LABEL/$h.aparc.annot','./$h.white-mask','$CRAS_PATH',ctx='$h',vn=$SURF_VN,add_lbl=[2,41])"
 done
 #Sub-cortical:
 for h in lh rh
 do
-python -c "import reconutils; reconutils.sample_vol_on_surf('$SURF/$h.aseg','./$vol-mask.nii.gz','$LABEL/$h.aseg.annot','./$h.aseg-mask','$CRAS_PATH',ctx=None,vn=$SURF_VN,add_lbl=[])"
+python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.sample_vol_on_surf('$SURF/$h.aseg','./$vol-mask.nii.gz','$LABEL/$h.aseg.annot','./$h.aseg-mask','$CRAS_PATH',ctx=None,vn=$SURF_VN,add_lbl=[])"
 done
 
 #Quality control snapshots:
@@ -266,7 +266,7 @@ then
     out_consim_path=./consim-vol-counts$STRMLNS_SIFT_NO-v$VOX.npy
     #Compute the voxel connectivity similarity
     #This takes a lot of time...
-    python -c "import reconutils; reconutils.node_connectivity_metric('$DMR/vol-counts$STRMLNS_SIFT_NO-v$VOX.npy',metric='cosine', mode='sim',out_consim_path='$out_consim_path')"
+    python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.node_connectivity_metric('$DMR/vol-counts$STRMLNS_SIFT_NO-v$VOX.npy',metric='cosine', mode='sim',out_consim_path='$out_consim_path')"
 
 else
     ref_vol_path=''
@@ -276,7 +276,7 @@ fi
 
 for h in lh rh
 do
-    python -c "import reconutils; import os; reconutils.connectivity_geodesic_subparc('$SURF/$h.white', '$LABEL/$h.aparc.annot', './$h.white-mask-idx.npy', out_annot_path='$LABEL/$h.aparc$SUBAPARC_AREA-$SUBAPARC_MODE.annot', parc_area=int('$SUBAPARC_AREA'), labels=None, hemi='$h', ctx='$h', mode='$SUBAPARC_MODE', cras_path='$CRAS_PATH', ref_vol_path='$ref_vol_path', consim_path='$out_consim_path', lut_path='$FREESURFER_HOME/FreeSurferColorLUT.txt')"
+    python -c "import bnm.recon.algo.reconutils; import os; bnm.recon.algo.reconutils.connectivity_geodesic_subparc('$SURF/$h.white', '$LABEL/$h.aparc.annot', './$h.white-mask-idx.npy', out_annot_path='$LABEL/$h.aparc$SUBAPARC_AREA-$SUBAPARC_MODE.annot', parc_area=int('$SUBAPARC_AREA'), labels=None, hemi='$h', ctx='$h', mode='$SUBAPARC_MODE', cras_path='$CRAS_PATH', ref_vol_path='$ref_vol_path', consim_path='$out_consim_path', lut_path='$FREESURFER_HOME/FreeSurferColorLUT.txt')"
 
     #Quality control snapshot
     python -m $SNAPSHOT --snapshot_name subaparc-aparc$SUBAPARC_AREA-$SUBAPARC_MODE-$h surf_annot $SURF/$h.inflated $LABEL/$h.aparc$SUBAPARC_AREA-$SUBAPARC_MODE.annot
@@ -285,7 +285,7 @@ done
 for h in lh rh
 do
     aseglist=ASEG_LIST_$h
-    python -c "import reconutils; import os; reconutils.connectivity_geodesic_subparc('$SURF/$h.aseg', '$LABEL/$h.aseg.annot', './$aseg-mask-idx.npy', out_annot_path='$LABEL/$h.aseg$SUBAPARC_AREA-$SUBAPARC_MODE.annot', parc_area=int('$SUBAPARC_AREA'), labels='${!aseglist}', hemi=None, ctx=None, mode='$SUBAPARC_MODE', cras_path='$CRAS_PATH', ref_vol_path='$ref_vol_path', consim_path='$out_consim_path', lut_path='$FREESURFER_HOME/FreeSurferColorLUT.txt')"
+    python -c "import bnm.recon.algo.reconutils; import os; bnm.recon.algo.reconutils.connectivity_geodesic_subparc('$SURF/$h.aseg', '$LABEL/$h.aseg.annot', './$aseg-mask-idx.npy', out_annot_path='$LABEL/$h.aseg$SUBAPARC_AREA-$SUBAPARC_MODE.annot', parc_area=int('$SUBAPARC_AREA'), labels='${!aseglist}', hemi=None, ctx=None, mode='$SUBAPARC_MODE', cras_path='$CRAS_PATH', ref_vol_path='$ref_vol_path', consim_path='$out_consim_path', lut_path='$FREESURFER_HOME/FreeSurferColorLUT.txt')"
 
     #Quality control snapshot
     python -m $SNAPSHOT --snapshot_name subaparc-aseg$SUBAPARC_AREA-$SUBAPARC_MODE-$h surf_annot $SURF/$h.aseg $LABEL/$h.aseg$SUBAPARC_AREA-$SUBAPARC_MODE.annot
