+++ bin/subaparc.sh
@@ -7,9 +7,9 @@ for area in $SUBAPARC_AREA
 do
     for h in lh rh
     do
-        python -c "import reconutils; reconutils.subparc_files('$h','aparc','aparc.sub$area',$area)"
+        python -c "import bnm.recon.algo.reconutils; bnm.recon.algo.reconutils.subparc_files('$h','aparc','aparc.sub$area',$area)"
         #or equivalently if PYTHONPATH is NOT set:
-        #python $CODE/reconutils.py subparc $h aparc aparc$area $area
+        #python bnm.recon.algo.reconutils.py subparc $h aparc aparc$area $area
     done
